Alantra python backend/requirements.txt
pydantic>=2.6,<3
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
supabase>=2.5,<3
python-dateutil>=2.9,<3
pytest>=8,<9
httpx>=0.27,<1
